editor_backend/requirements.txt
Django==4.2.5
django-cors-headers==4.2.0
docx2txt==0.8